// src/api/generated/schemas/apiKeyResponse.ts
/**
 * Generated by orval v7.12.2 🍺
 * Do not edit manually.
 * Notifiable API
 * A simple example ASP.NET Core Web API
 * OpenAPI spec version: v1
 */

export interface ApiKeyResponse {
  id?: string;
  name?: string;
  keyPrefix?: string;
  scopes?: string[];
  createdAt?: string;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  revokedAt?: string | null;
  /** @nullable */
  lastUsedAt?: string | null;
  isActive?: boolean;
}